requirements.txt
google-cloud-storage
google-auth==1.6.2
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-cloud-pubsub==0.39.1
paho-mqtt==1.4.0
pyjwt==1.7.1
oauth2client